rename utils package to sdk
