+++ aito/cli/sub_commands/convert.py
@@ -2,8 +2,8 @@ import json
 import sys
 from typing import Dict, List
 
-from aito.utils.data_frame_handler import DataFrameHandler
-from aito.utils.schema_handler import SchemaHandler
+from aito.sdk.data_frame_handler import DataFrameHandler
+from aito.sdk.schema_handler import SchemaHandler
 from .sub_command import SubCommand
 from ..parser import PathArgType, InputArgType, ParseError
 from ..parser_utils import try_json_load
