+++ aito/cli/sub_commands/infer_table_schema.py
@@ -3,8 +3,8 @@ import sys
 from typing import Dict
 from typing import List
 
-from aito.utils.data_frame_handler import DataFrameHandler
-from aito.utils.schema_handler import SchemaHandler
+from aito.sdk.data_frame_handler import DataFrameHandler
+from aito.sdk.schema_handler import SchemaHandler
 from .sub_command import SubCommand
 from ..parser import InputArgType, ParseError
 from ..parser_utils import create_sql_connecting_from_parsed_args
