+++ aito/common/generic_utils.py
