+++ tests/cli/test_parser_utils.py
@@ -1,6 +1,6 @@
 from aito.cli.parser import ParseError, ArgParser
 from aito.cli.parser_utils import parse_env_variable, create_client_from_parsed_args
-from aito.utils.aito_client import AitoClient, BaseError
+from aito.sdk.aito_client import AitoClient, BaseError
 from tests.cases import CompareTestCase
