+++ tests/sdk/test_data_frame_handler.py
@@ -2,7 +2,7 @@ import json
 
 import ndjson
 
-from aito.utils.data_frame_handler import DataFrameHandler
+from aito.sdk.data_frame_handler import DataFrameHandler
 from tests.cases import CompareTestCase
